+++ src/Control/Row.php
@@ -1,0 +1,11 @@
+<?php
+
+namespace Efabrica\NeoForms\Control;
+
+class Row extends Div
+{
+    public function __construct()
+    {
+        parent::__construct('row');
+    }
+}